feat: get hashtag post adjusts
